// frontend/lms/src/api/models/BookUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BookUpdate = {
    title?: (string | null);
    description?: (string | null);
    price?: (number | null);
    quantity?: (number | null);
    page_count?: (number | null);
    dimensions?: (string | null);
    weight?: (number | null);
    table_of_contents?: (string | null);
    introduction?: (string | null);
    publisher_image?: (string | null);
    cover_image?: (string | null);
    author_id?: (string | null);
    isbn?: (string | null);
};